refactor(api-subscribe): move component under api
fix gravitee-io/issues#3302
